requirements.txt
anthropic==0.49.0
fal-client==0.5.6
Pillow==10.3.0
gspread==6.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0
python-dotenv==1.0.1
requests==2.31.0
klaviyo-api==10.0.0
